Revert "move sampler test to one file"
This reverts commit b80ad25ebd9df27d1d8adaeda938d4569a413fe9.
